core/ratelimit bugfix: race can lead to segfault
There was a race in iminternalAddMsg(), where the mutex is
released and after that the passed message object is accessed.
If the mainloop iterates in between, the msg may have already
been deleted by this time, resulting in a segfault.

Most importantly, there is no need to release the mutex lock
early, as suggested by current code. Inside mainloop the mutex
is acquired when it is time to do so, so at worst we'll have a
very slight wait there (which really does not matter at all).

This only happens if a large number of internal messages are emitted.

closes https://github.com/rsyslog/rsyslog/issues/1828

(cherry picked from commit ba5784b635ee1e7dbf7e33d75eb12b02fd44afca)